mapify_project/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.2
pymysql==1.0.2
Werkzeug==2.0.2
transformers==4.18.0
torch==1.11.0
spacy==3.5.0
networkx==2.8.4
matplotlib==3.5.2